local_requirements.txt
# Additional packages useful for development
django-extensions
ipython
django-debug-toolbar      # debug
pytest-django==3.8.0      # test
pytest==5.3.5             # via pytest-django
coverage                  # test coverage



# for ads maker

site-map-parser
lxml==4.4.1
#pytest-cov==2.8.1
#requests-mock==1.7.0
python-dateutil==2.8.1

requests # requests!!!
beautifulsoup4 # beautifulsoop
